casting-defect-computer-vision/casting-defect-detection/requirements.txt
kelvin-python-sdk
numpy
tensorflow
opencv-python
pillow